refactor: rearranged code focused on updates to flow better
The original way I was triggering processes caused some issues, the primary one was I coupled the server starting with updating. This prevented there being clean way to separate the task of just restarting the server and updating the server, which was the goal of the current feature. Now the ark bootstrap performs checks for initial server download and if it should be auto updated. It either will launch the ark updater or just launch the server. The updater now launches the server after it has finished. This removed the need for the complex status file checking, because the ark server will not launch if the update/validation fails. Overall this should reduce the complexity of the code and also decouple the core processes.

ark-sa-updater
* added launch_ark_sa_server which is called after the update is finished
* removed all status file code

ark-sa-server
* removed all code based around updates and the updater status file
* refined how launch flags were aggregated and moved to local vars

ark-sa-bootstrap
* added check_if_server_files_exists which launches updater if needed
* added launch_ark_sa_server which launches the server if no downloads/updated needed
* added launch_update_service which is responsible for the supervisorctl launch of updater
* added auto_update_server which will launch updater if auto update flag is set

docekrfile
* removed UPDATE_TIMEOUT env var